* void SUDO by using "--dev null" instead of tun * some tweaks to allow openvpn.exe unittesting for {udp,tcp6}-remote

+++ run-all.sh
@@ -9,9 +9,17 @@ source ${0%/*}/my_env.sh || exit 1
 source ${0%/*}/libtest.sh || exit 1
 
 t=$LIBTEST_OUTPUT_DIR
-cp -pu $PWD/openvpn $t/openvpn-test || exit 1
-my_dir=${0%/*}
-export OPENVPN="$t/openvpn-test"
+tdir=${0%/*}
+
+# no-op, just to document :)
+: ${O_ARGS:=}
+if [ -f openvpn ];then
+  cp -pu $PWD/openvpn $t/openvpn-test || exit 1
+  export OPENVPN="$t/openvpn-test"
+elif [ -f openvpn.exe ];then
+  cp -pu $PWD/openvpn.exe $t/openvpn-test.exe || exit 1
+  export OPENVPN="$t/openvpn-test.exe"
+fi
 
 all_tests() {
 local post="$1"
@@ -21,44 +29,36 @@ test_set_cleanup "ps ax | awk '/[o]penvpn-test/ { print \$1 }'  | ${SUDO} xargs 
 trap 'test_bg_cleanup;exit' 0 2 15
 
 test_define "UDP6 loopback$post"
-test_bg_egrep 30 "Initialization Sequence Completed" ${my_dir?}/run-udp6-0-loopback.sh
+test_bg_egrep 30 "Initialization Sequence Completed" ${tdir?}/run-udp6-0-loopback.sh $O_ARGS
 
 test_define "UDP6 loopback byname$post"
-test_bg_egrep 30 "Initialization Sequence Completed" ${my_dir?}/run-udp6-0-loopback-byname.sh
+test_bg_egrep 30 "Initialization Sequence Completed" ${tdir?}/run-udp6-0-loopback-byname.sh $O_ARGS
 
 if ping6 -c 1 "$REM6" >/dev/null 2>&1;then
-  SUDO=sudo
-  test_bg_cleanup ## special SUDO forces
   test_define "UDP6 remote$post"
-  test_bg_prev ${my_dir?}/run-udp6-1-rem.sh --ping-exit 30
-  test_bg_egrep 30 "Initialization Sequence Completed" ${my_dir?}/run-udp6-1-loc.sh --ping-exit 30
+  test_bg_prev ${tdir?}/run-udp6-1-rem.sh --ping-exit 60 $O_ARGS
+  test_bg_egrep 30 "Initialization Sequence Completed" ${tdir?}/run-udp6-1-loc.sh --ping-exit 60 $O_ARGS
 
-  test_bg_cleanup ## special SUDO forces
-  sudo killall -q -9 openvpn-test
   test_define "TCP6 remote$post"
-  test_bg_prev ${my_dir?}/run-tcp6-1-server.sh --ping-exit 30
-  sleep 10
-  test_bg_egrep 30 "Initialization Sequence Completed" ${my_dir?}/run-tcp6-1-client.sh --ping-exit 30
-  sudo killall -q -9 openvpn-test
-  test_bg_cleanup ## special SUDO forces
-  unset SUDO
+  test_bg_prev ${tdir?}/run-tcp6-1-server.sh --ping-exit 60
+  test_bg_egrep 30 "Initialization Sequence Completed" ${tdir?}/run-tcp6-1-client.sh --ping-exit 60 $O_ARGS
 else
   notice "REM6=\"$REM6\" not reachable, skipping remote UDP6,TCP6 tests"
 fi
 test_define "TCP6 loopback$post"
-test_bg_prev ${my_dir?}/run-tcp6-0-loopback-server.sh
-test_bg_egrep 30 "Initialization Sequence Completed" ${my_dir?}/run-tcp6-0-loopback-client.sh
+test_bg_prev ${tdir?}/run-tcp6-0-loopback-server.sh $O_ARGS
+test_bg_egrep 30 "Initialization Sequence Completed" ${tdir?}/run-tcp6-0-loopback-client.sh $O_ARGS
 
 test_define "TCP6 loopback byname$post"
-test_bg_prev ${my_dir?}/run-tcp6-0-loopback-server-byname.sh
-test_bg_egrep 30 "Initialization Sequence Completed" ${my_dir?}/run-tcp6-0-loopback-client-byname.sh
+test_bg_prev ${tdir?}/run-tcp6-0-loopback-server-byname.sh $O_ARGS
+test_bg_egrep 30 "Initialization Sequence Completed" ${tdir?}/run-tcp6-0-loopback-client-byname.sh $O_ARGS
 
 test_define "UDP4 loopback$post"
-test_bg_egrep 30 "Initialization Sequence Completed" ${my_dir?}/run-udp4-0-loopback.sh
+test_bg_egrep 30 "Initialization Sequence Completed" ${tdir?}/run-udp4-0-loopback.sh $O_ARGS
 
 test_define "TCP4 loopback$post"
-test_bg_prev ${my_dir?}/run-tcp4-0-loopback-server.sh 
-test_bg_egrep 30 "Initialization Sequence Completed" ${my_dir?}/run-tcp4-0-loopback-client.sh
+test_bg_prev ${tdir?}/run-tcp4-0-loopback-server.sh  $O_ARGS
+test_bg_egrep 30 "Initialization Sequence Completed" ${tdir?}/run-tcp4-0-loopback-client.sh $O_ARGS
 
 }
